Don't use 'podman cp' to copy toolbox.sh to old containers
Copying files into a running container is considered inherently hacky.
Rootful Podman can pause a container using 'podman cp --pause ...'
during the copy, but that's not possible when used rootless.

Secondly 'podman cp' has suffered from a series of regressions lately.
First there was the problem with how the --pause flag was handled [1],
and then there's an unsolved regression where /etc/profile.d/toolbox.sh
is being created as a directory by:
  $ podman cp \
            --pause=false \
            /etc/profile.d/toolbox.sh \
            "$container":/etc/profile.d

Try to side-step all that by using $XDG_RUNTIME_DIR as a conduit to
share the file with the container and using plain cp(1) to place it in
the toolbox container's /etc/profile.d.

[1] Commit e715ff2f9b231271
    https://github.com/debarshiray/toolbox/pull/193

https://github.com/debarshiray/toolbox/issues/196